Fix passed over count
At some point after writing the code that populated
'passed_over_urls' and writing the code that used 'passed_over_urls'
I forgot what this column actually was

It includes every examined URL except the final click.
So it includes things that were clicked.

This means all the code is wrong, because it assumed
these were only the not-clicked ones.

The reason for storing all the URLs was so that the dataset
includes the rank of each examined document.