SDBN using pyclick
My existing SDBN implementation generates some dodgy results.
The calculation itself doesn't take into account the amount of
observed data, so it gets overconfident about particular
results that are currently ranked very low.

I tried to account for this with a separate uncertainty calculation,
but I don't think this is sound.

In Chapelle & Zhang (2009) they use a beta(1,1) prior which means that
in the absense of data attractiveness/satisfyingness params are assumed
to be close to 0.5. This may not be a good description of prior
knowledge but it's less distruptive than ignoring it entirely.

In their paper they also describe how to calculate a confidence
for each of the estimated parameters, and I would like to be able
to take this into account when producing a recommended ranking
from the model.

However, the full DBN model is much more complicated to implement
than the SDBN, so I want to use PyClick's existing implementation
to see how different the models are before I spend a lot of time
trying to understand how the expectation maximisation steps work.

This commit is just the SDBN.

For my current dataset (queries with >1000 searches over a few weeks)
this gives:

-------------------------------
Testing on 103312 search sessions (127 unique queries).
-------------------------------
  log-likelihood: -0.068486; time: 12 secs
  perplexity: 1.104745; time: 14 secs